:truck: Renamed folder names and changed paths accordingly and created new folders as well
