+++ .env
@@ -1,0 +1,1 @@
+VITE_API_BASE_URL = "https://reqres.in/api/"
