+++ src/App.jsx
@@ -1,14 +1,14 @@
 import AppStyle from "@src/styled/App.style";
-import Container from "@src/container/Container";
-import UserList from "@src/components/UserList/UserList";
+import UserListContainer from "@src/containers/UserListContainer";
+import UserListWrap from "@src/components/UserList/UserListWrap";
 
 function App() {
   return (
     <>
       <AppStyle>
-        <Container>
-          <UserList />
-        </Container>
+        <UserListContainer>
+          <UserListWrap />
+        </UserListContainer>
       </AppStyle>
     </>
   );
