+++ src/components/Navbar/index.js
@@ -1,3 +1,3 @@
 import Navbar from "./Navbar";
 
-export {Navbar};
+export { Navbar };
