+++ src/components/NavigationButton/index.js
@@ -1,0 +1,3 @@
+import Button from "./Button";
+
+export { Button };
