+++ src/components/UserList/index.js
@@ -1,0 +1,3 @@
+import UserListDropDown from "@src/DropDown/UserListDropDown";
+
+export default UserListDropDown;
