--- src/components/UserList/index.jsx
@@ -1,3 +1,0 @@
-import DropDown from "./DropDown";
-
-export default DropDown;
