+++ src/config/env.js
@@ -1,0 +1,1 @@
+export const baseURL = import.meta.env.VITE_API_BASE_URL;
